fix: #1033 QAレビュー指摘修正 (SSOT違反・checkAuth共通化・制限値ハードコード)
- tierLabel を getPlanShortLabel() 経由に修正 (用語辞書 SSOT 準拠)
- checkCronAuth を $lib/server/auth/cron-auth.ts に共通化 (2箇所のコピペ解消)
- メール内のプラン制限値を plan-limit-service の getPlanLimits('free') から取得
- DynamoDB findActiveTrials スタブにスコープ外コメントを追記
